Move BerthOffer to its own dir & extract BerthApplication
